fix non-luadata changes not persisting
It seems that all save game data that isn't part of vars is saved before
BeforeNewGameAutosave and BeforeSaveGame are run. To work around this,
in each case I added a just set a flag indicating that something needs
to be initialized and delay initializing until BeforeLoadMap (and clear
the flag).